Subject: gcc format strings: Make GCC check format strings, fix found problems

+++ test-server/test-server-echogen.c
@@ -107,8 +107,8 @@ callback_lws_echogen(struct lws *wsi, enum lws_callback_reasons reason,
 		break;
 
 	case LWS_CALLBACK_WS_PEER_INITIATED_CLOSE:
-		lwsl_notice("LWS_CALLBACK_WS_PEER_INITIATED_CLOSE: len %d\n",
-			    len);
+		lwsl_notice("LWS_CALLBACK_WS_PEER_INITIATED_CLOSE: len %lu\n",
+			    (unsigned long)len);
 		for (n = 0; n < (int)len; n++)
 			lwsl_notice(" %d: 0x%02X\n", n,
 				    ((unsigned char *)in)[n]);
